client-hints/accept_ch_lifetime-manual.https.html

<html>
<title>Accept-CH-Lifetime manual test</title>
<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<div id=test>
    <p>This is a manual test, since the tests requires the browser to
    have a cache implementation for storing client hints preferences of
    different origins. Such a cache implementation may only be available
    only when the full implementation of browser is run.

    The test additionally opens another html web page. One test is ran in this
    web page, and another in the second web page.
</div>

<script>

// This test fetches resources/accept_ch_lifetime.html. The response headers to
// that webpage contain Accept-CH and Accept-CH-Lifetime headers.
// Fetching that webpage should cause the user-agent to persist origin
// preferences for the client hints specified in Accept-CH header for a
// duration specified in the Accept-CH-Lifetime header.

// Next, to verify if the origin preferences were persisted by the user
// agent, this test fetches resources/expect_client_hints_headers.html
// in a new window. Fetching of resources/expect_client_hints_headers.html
// verifies that the user agent actually sends the client hints in the request
// headers.

promise_test(t => {
  return fetch("resources/accept_ch_lifetime.html", {"mode": "no-cors"}).then(r => {
    assert_equals(r.status, 200)
    assert_false(r.headers.has("device-memory-received"), "device-memory-received");
  });
}, "Test receiving Accept-CH-Lifetime header");

// Open a new window.
window.open("resources/expect_client_hints_headers.html");

</script>

</body>
</html>
